Smarter way to load total distance from UserDefaults
Distance is being measured sometimes even if the app is in background (untill it is suspended). User defaults are stored as soon as app goes to the background, and when it is loaded (app in foreground) all background changes are lost. Now we check if actuall distance is greater than the one stored in UserDefaults and if so, just use current distance.

Also small refactor for the rest of the code.